Users Endpoints added with exception

+++ app/crud/posts_crud.py
@@ -1,17 +1,17 @@
 from app.database import get_session
 from app.model import UPosts
-from sqlmodel import select, Session
+from sqlmodel import select
 
 session = next(get_session())
 
-def insert_data_in_table(post: UPosts):
+def insert_data_in_uposts_table(post: UPosts):
     session.add(post)
     session.commit()
     session.refresh(post)
     session.close()
     return post
 
-def select_all_post():    
+def get_all_post():    
         posts = session.exec(select(UPosts)).all()
         return posts
 
@@ -31,7 +31,8 @@ def update_post_by_id(id: int, upost: UPosts):
 
 def delete_post_by_id(id: int):
     upost = session.get(UPosts, id)
-    session.delete(upost)
-    session.commit()
+    if upost:
+        session.delete(upost)
+        session.commit()
     return upost
